backend/requirements.txt
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
alembic==1.12.1
pytest==7.4.3
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0
